backend/pyproject.toml
[project]
name = "auraclass-backend"
version = "0.1.0"
description = "AuraClass班级量化管理系统后端"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.1",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "casbin>=1.24.0",
    "casbin-sqlalchemy-adapter>=0.4.1",
    "pydantic>=2.0.0",
    "pydantic_settings>=2.8.1",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "redis>=4.6.0",
    "httpx>=0.24.1",
    "prometheus-client>=0.16.0",
    "psutil>=5.9.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "pylint>=2.17.4"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
plugins = [
    "pydantic.mypy"
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88